Move LoginPage to a separate package and make BasePage parent class of LoginPage
