+++ src/main/java/ru/siobko/testing/tasks/junit/tests/GroupCreationTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import ru.siobko.testing.tasks.junit.core.main.FeedPage;
 import ru.siobko.testing.tasks.junit.core.main.GroupPage;
 import ru.siobko.testing.tasks.junit.core.main.GroupsPage;
-import ru.siobko.testing.tasks.junit.core.main.LoginPage;
+import ru.siobko.testing.tasks.junit.core.login.LoginPage;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -15,17 +15,13 @@ public class GroupCreationTest extends BaseTest {
 
     @BeforeAll
     public static void setUp() {
-        LoginPage loginPage = new LoginPage();
-        loginPage.login(EMAIL, PASSWORD);
-
-        FeedPage feedPage = new FeedPage();
+        FeedPage feedPage = new LoginPage().login(EMAIL, PASSWORD);
         feedPage.openGroupsPage();
     }
 
     @Test
     public void testGroupCreation() {
-        GroupsPage groupsPage = new GroupsPage();
-        GroupPage groupPage = groupsPage.createGroup(GROUP_NAME);
+        GroupPage groupPage = new GroupsPage().createGroup(GROUP_NAME);
 
         assertTrue(
                 groupPage.checkGroupNameEquals(GROUP_NAME),
@@ -35,7 +31,6 @@ public class GroupCreationTest extends BaseTest {
 
     @AfterAll
     public static void tearDown() {
-        GroupPage groupPage = new GroupPage();
-        groupPage.deleteGroup();
+        new GroupPage().deleteGroup();
     }
 }
