+++ src/main/java/ru/siobko/testing/tasks/junit/tests/LoginTest.java
@@ -1,7 +1,7 @@
 package ru.siobko.testing.tasks.junit.tests;
 
 import org.junit.jupiter.api.Test;
-import ru.siobko.testing.tasks.junit.core.main.LoginPage;
+import ru.siobko.testing.tasks.junit.core.login.LoginPage;
 import ru.siobko.testing.tasks.junit.core.main.FeedPage;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
