+++ src/main/java/ru/siobko/testing/tasks/junit/tests/LogoutTest.java
@@ -3,21 +3,19 @@ package ru.siobko.testing.tasks.junit.tests;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import ru.siobko.testing.tasks.junit.core.main.FeedPage;
-import ru.siobko.testing.tasks.junit.core.main.LoginPage;
+import ru.siobko.testing.tasks.junit.core.login.LoginPage;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class LogoutTest extends BaseTest {
     @BeforeAll
     public static void setUp() {
-        LoginPage loginPage = new LoginPage();
-        loginPage.login(EMAIL, PASSWORD);
+        new LoginPage().login(EMAIL, PASSWORD);
     }
 
     @Test
     public void testLogin() {
-        FeedPage feedPage = new FeedPage();
-        feedPage.openUserCardToolbar()
+        new FeedPage().openUserCardToolbar()
                 .logout()
                 .confirmLogout();
